Refactor: TeamCard 리스트 형식의 컴포넌트 분리
