+++ src/components/common/TeamCardsList.scss
@@ -1,0 +1,7 @@
+.teamcards {
+  margin-top: 80px;
+  width: 1200px;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr 1fr;
+  gap: 20px;
+}
