+++ src/components/signup/SignupStepOne.scss
@@ -11,11 +11,4 @@
     margin-top: 25px;
     color: $login-text;
   }
-  .teamcards-block {
-    margin-top: 80px;
-    width: 1200px;
-    display: flex;
-    flex-wrap: wrap;
-    gap: 20px;
-  }
 }
